Preliminary support for profiling HaLVMs.
Profiling is supported by wiring the various stdio calls that GHC uses for
profiling into an extremely impoverished disk driver implemented in
xen/profiling.c. Note that the whole framework is a bit fragile, with a
lot of hard-coded assumptions and constants, like the disks to use for
profiling data and the sector size of those disks. These might be worth
fixing up later.